add expense, cleanup/refactoring
